backend/app/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0.1
pydantic>=2.7
SQLAlchemy>=2.0
psycopg2-binary>=2.9
langchain>=0.3
langchain-core>=0.3
langchain-community>=0.3
langchain-google-genai>=2.0.0
google-generativeai>=0.7.2
faiss-cpu>=1.8.0.post1
PyMuPDF>=1.24
python-dateutil>=2.9.0.post0
httpx>=0.27
    